+++ ufotest/util.py
@@ -6,8 +6,9 @@ import click
 import random
 import re
 import subprocess
+import shutil
 import importlib.util
-from typing import Optional, Tuple, Dict
+from typing import Optional, Tuple, Dict, List
 from abc import ABC, abstractmethod
 
 from jinja2 import Template, FileSystemLoader, Environment